Refactor polling components (namespace/namings)
